+++ tests/dummy/app/controllers/example-2.js
@@ -1,3 +1,5 @@
 import ThingsController from '../_controllers/things';
 
-export default ThingsController;
+export default class Example2Controller extends ThingsController {
+  document = document;
+}