+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 8.0.0-beta.1
+
+- Removes `@selector` in favour of `@element`
+- Removes `@useDocument` in favour of `@element`
+- Adds `scroller.setElement` to make setting child elements easier
+- Renames `@scrollDebounce` to `@debounce`
+
 ## 7.0.7
 
 - Fix division by zero